requirements.txt
# Core Dependencies
smolagents>=0.1.0
python-dotenv>=1.0.0

# LLM Integrations
openai>=1.0.0
anthropic>=0.5.0
transformers>=4.30.0
torch>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
requests>=2.30.0

# API Integrations
google-search-results>=2.4.2  # For SerpAPI
brave-search>=1.0.0  # For Brave Search API

# Data Visualization
rich>=13.4.0
terminaltables>=3.1.10
tabulate>=0.9.0
termgraph>=0.5.3
termplotlib>=0.4.0

# Database
SQLAlchemy>=2.0.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.65.0
validators>=0.20.0
langchain>=0.0.235  # Optional for some components
